Match PrepareBattleGraphicsMaybe

+++ include/anime.h
@@ -196,7 +196,7 @@ void AnimDisplay(struct Anim * anim);
 bool PrepareBattleGraphicsMaybe(void);
 u16 sub_8057CAC(struct Unit * unit, const struct BattleAnimDef * pBattleAnimDef, u16, int * out);
 int GetBanimTerrainGround(u16 terrain, u16 tileset);
-s8 GetBanimBackgroundIndex(u16 terrain, u16 tileset);
+int GetBanimBackgroundIndex(u16 terrain, u16 tileset);
 s16 GetSpellAnimId(u16 jid, u16 weapon);
 void UnsetMapStaffAnim(s16 * out, u16 pos, u16 weapon);
 void ParseBattleHitToBanimCmd(void);
